% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanicon.R
\name{dependencies}
\alias{dependencies}
\alias{use_titanicon}
\alias{titanicon_opts}
\alias{titanicon_theme}
\alias{reset_titanicon_theme}
\alias{titanicon}
\title{Titanicon}
\usage{
use_titanicon()

titanicon_opts(hover = TRUE, click = TRUE)

titanicon_theme(...)

reset_titanicon_theme()

titanicon(icon, ..., class = "")
}
\arguments{
\item{hover, click}{Whether to trigger the animation on hover and click.}

\item{...}{Any other option or element passed to \link[shiny]{div}.}

\item{icon}{An icon, see below for valid icons.}

\item{class}{Any additional class(es).}
}
\description{
All titanicon functions.
}
\section{Icons}{

\itemize{
  \item{\code{caps}}
  \item{\code{chat}}
  \item{\code{checkbox}}
  \item{\code{expand}}
  \item{\code{cheap}}
  \item{\code{expensive}}
  \item{\code{idea}}
  \item{\code{mailbox}}
  \item{\code{mic}}
  \item{\code{no-mic}}
  \item{\code{online}}
  \item{\code{pause}}
  \item{\code{power}}
  \item{\code{shopping}}
  \item{\code{smile}}
  \item{\code{stop}}
  \item{\code{unlock}}
  \item{\code{zoom}}
}
}

